// src/shared/models/qod.tsx
// Generated by https://quicktype.io

export interface QuoteOfDay {
  success?: Success
  contents?: Contents
  error?: string
}

export interface Contents {
  quotes: Quote[]
  copyright: string
}

export interface Quote {
  quote: string
  length: string
  author: string
  tags: string[]
  category: string
  date: string
  permalink: string
  title: string
  background: string
  id: string
}

export interface Success {
  total: number
}